task-management-system/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
filelock==3.16.1
email-validator==2.2.0

motor  #mongodb(async
asyncpg  #postgresql
aiomysql #mysql(async)
sqlalchemy
python-dotenv